requirements.txt
aiohttp==3.12.13
aiogram==3.21.0
redis==5.2.0
python-dotenv==1.1.1
aiohttp==3.12.13
asyncpg==0.29.0
APScheduler==3.11.0
loguru
requests==2.32.4
httpx